Fix Dao return type and create alias for query
