+++ app/src/main/java/com/example/fitnesslog/exercise/domain/model/SetTemplate.kt
@@ -3,7 +3,6 @@ package com.example.fitnesslog.exercise.domain.model
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.ForeignKey
-import androidx.room.Ignore
 import androidx.room.Index
 import androidx.room.PrimaryKey
 
@@ -31,6 +30,4 @@ data class SetTemplate(
     @ColumnInfo(name = "position") val position: Int,
     @ColumnInfo(name = "created_at") val createdAt: Long,
     @ColumnInfo(name = "updated_at") val updatedAt: Long,
-    @Ignore var isDirty: Boolean = false,
-    @Ignore var isDeleted: Boolean = false
 )
