+++ app/src/main/java/com/example/fitnesslog/history/ui/history/fragment/HistoryFragment.kt
@@ -1,4 +1,4 @@
-package com.example.fitnesslog.program.view.fragment
+package com.example.fitnesslog.history.ui.history.fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -14,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [ProgramFragment.newInstance] factory method to
+ * Use the [HistoryFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class ProgramFragment : Fragment() {
+class HistoryFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -35,21 +35,7 @@ class ProgramFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_program, container, false)
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-//        val programs = ProgramDataSource().loadPrograms()
-//        val programAdapter = ProgramAdapter(programs)
-//
-//        val rvPrograms: RecyclerView = view.findViewById(R.id.rvPrograms)
-//        rvPrograms.adapter = programAdapter
-//        rvPrograms.layoutManager = GridLayoutManager(context, 2)
-//        rvPrograms.addItemDecoration(GridSpacingItemDecoration(20))
-
-
+        return inflater.inflate(R.layout.fragment_history, container, false)
     }
 
     companion object {
@@ -59,12 +45,12 @@ class ProgramFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment ProgramFragment.
+         * @return A new instance of fragment HistoryFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            ProgramFragment().apply {
+            HistoryFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
