+++ app/src/main/java/com/example/fitnesslog/program/ui/program_edit/adapter/ProgramEditAdapter.kt
@@ -1,0 +1,4 @@
+package com.example.fitnesslog.program.ui.program_edit.adapter
+
+class ProgramEditAdapter {
+}
