--- app/src/main/java/com/example/fitnesslog/program/view/adapter/ProgramEditAdapter.kt
@@ -1,4 +1,0 @@
-package com.example.fitnesslog.program.view.adapter
-
-class ProgramEditAdapter {
-}
