+++ app/src/main/res/layout/fragment_program.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".program.view.fragment.ProgramFragment">
+    tools:context=".program.ui.program.fragment.ProgramFragment">
 
     <TextView
         android:id="@+id/tvProgramFragmentTitle"
