+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/historyFragment"
-        android:name="com.example.fitnesslog.history.view.fragment.HistoryFragment"
+        android:name="com.example.fitnesslog.history.ui.history.fragment.HistoryFragment"
         android:label="@string/label_history"
         tools:layout="@layout/fragment_history">
         <action
@@ -19,7 +19,7 @@
     </fragment>
     <fragment
         android:id="@+id/workoutHomeFragment"
-        android:name="com.example.fitnesslog.workout.view.fragment.WorkoutHomeFragment"
+        android:name="com.example.fitnesslog.workout.ui.workout_home.fragment.WorkoutHomeFragment"
         android:label="@string/label_workout"
         tools:layout="@layout/fragment_workout_home">
         <action
@@ -31,7 +31,7 @@
     </fragment>
     <fragment
         android:id="@+id/programFragment"
-        android:name="com.example.fitnesslog.program.view.fragment.ProgramFragment"
+        android:name="com.example.fitnesslog.program.ui.program.fragment.ProgramFragment"
         android:label="@string/label_program"
         tools:layout="@layout/fragment_program">
         <action
